requirements.txt
accelerate==0.15.0
certifi==2022.12.7
charset-normalizer==2.1.1
diffusers==0.11.1
filelock==3.9.0
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==6.0.0
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==22.0
Pillow==9.4.0
psutil==5.9.4
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
scipy==1.10.0
tokenizers==0.13.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.25.1
typing-extensions==4.4.0
urllib3==1.26.13
zipp==3.11.0
